# swagger/v1/swagger.yaml
definitions:
...
    ErrorResponse:
        properties:
            code:
                type: string
                x-go-name: Code
            message:
                type: string
                x-go-name: Message
            severity:
                type: string
                x-go-name: Severity
            title:
                type: string
                x-go-name: Title
        type: object
        x-go-package: github.com/techx/portal/handler/response
...
paths:
...
    /public/user/profile:
        get:
            operationId: userProfile
            parameters:
                - in: header
                  name: X-User-Id
                  required: true
                  type: string
                  x-go-name: UserUUID
            responses:
                "200":
                    $ref: '#/responses/UserProfileResponse'
                "400":
                    description: ErrorResponse
                    schema:
                        $ref: '#/definitions/ErrorResponse'
                "401":
                    description: ""
                "422":
                    description: ErrorResponse
                    schema:
                        $ref: '#/definitions/ErrorResponse'
                "500":
                    description: ErrorResponse
                    schema:
                        $ref: '#/definitions/ErrorResponse'
                "503":
                    description: ErrorResponse
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            tags:
                - public
...
swagger: "2.0"